# Optimize get non-pool size