```
BenchmarkBeforeSmall			20000000	        51.0 ns/op
BenchmarkAfterSmall			30000000	        37.4 ns/op
BenchmarkBeforeNormal			10000000	       209 ns/op
BenchmarkAfterNormal			10000000	       210 ns/op
BenchmarkBeforeBig			10000000	       153 ns/op
BenchmarkAfterBig			10000000	       131 ns/op
```